Refactor header into smaller modules and drop duplicate sweden key.
Split navigation config, sub-menus, and desktop/mobile views into separate files under layout/header, and use header.nation instead of header.sweden.
