+++ src/components/layout/header/navStyles.ts
@@ -1,0 +1,5 @@
+export const NAV_TITLE_CLASS =
+  "block rounded-md px-2 py-1.5 text-sm font-medium text-white hover:bg-black-1 transition-colors";
+
+export const NAV_SUB_ITEM_CLASS =
+  "block rounded-md py-1 text-sm text-grey hover:bg-black-1 hover:text-white transition-colors";
