+++ src/locales/en/translation.json
@@ -13,7 +13,6 @@
     "explore": "Explore",
     "municipalities": "Municipalities",
     "nation": "Sweden",
-    "sweden": "Sweden",
     "regions": "Regions",
     "about": "About",
     "aboutUs": "About Us",
