+++ src/locales/sv/translation.json
@@ -13,7 +13,6 @@
     "explore": "Utforska",
     "municipalities": "Kommuner",
     "nation": "Sverige",
-    "sweden": "Sverige",
     "regions": "Län",
     "about": "Om",
     "aboutUs": "Om oss",
